% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{bound}
\alias{bound}
\title{returns the x-value in dens that is closest to the given x}
\usage{
bound(x, dens, return.x = TRUE)
}
\description{
returns the x-value in dens that is closest to the given x
}
