% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorUnderCurve.R
\name{color.den}
\alias{color.den}
\title{color area under a density within an interval
den has to be a density object}
\usage{
color.den(den, from, to, col.den = "white", col.area = "red", add = F,
  ...)
}
\description{
color area under a density within an interval
den has to be a density object
}
