% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd_mult.R
\name{hpd_mult}
\alias{hpd_mult}
\title{hpd_mult (by Mickey Warner)}
\usage{
hpd_mult(x, dens, prob = 0.95, tol, interactive = FALSE)
}
\arguments{
\item{x}{vector of samples from some distribution}

\item{dens}{a density object based on x, defaults to density(x)}

\item{prob}{numeric in (0, 1) the probability that the interval should be}

\item{tol}{numeric, smaller means longer compute time, but more accurate
results}

\item{interactive}{logical, defaults to FALSE. If TRUE, a fun plot is shown
and the function's process of choosing the region is displayed.
The user hits enter to proceed through.}
}
\description{
Compute the highest density posterior region from a (possibly multi-modal)
sample of points.
}
\details{
Since hpd_mult relies on a density object, it is possible to be returned a
result that is outside the bounds of the data (i.e., an hpd for a beta random
variable that has a left end point below 0).

A random horizontal line is repeatedly generated and the points along the line
that cross dens$y are calculated. These points make up a proposed hpd region
whose area is computed. Given this area, a new horizontal line is generated
until the area is close to prob (within tol).
}
\examples{
set.seed(1)
x = c(rnorm(100), rnorm(100, 5)
hpd = hpd_mult(x)
}
\seealso{
hpd_uni
}

