% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypairs.R
\name{my.pairs}
\alias{my.pairs}
\title{my paris function (prettier)}
\usage{
my.pairs(M, digits = 3, customDiag = NULL, customLower = NULL,
  customUpper = NULL)
}
\description{
my paris function (prettier)
}
\examples{
X <- matrix(rnorm(100),ncol=4)
colnames(X) <- c("I","II","III","IV")
my.pairs(X)
}
