% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInPlot.R
\name{plotInPlot}
\alias{plotInPlot}
\title{stay: acts like 'add' in the 'plot' function. The default is FALSE.}
\usage{
plotInPlot(minor.plot, coords = "topright", scale = 1/3, stay = F,
  bg.col = rgb(0, 0, 0, 0))
}
\arguments{
\item{minor.plot}{is a function with no parameters that plots the smaller plot}

\item{coords}{= x1,y1,x2,y2}
}
\description{
stay: acts like 'add' in the 'plot' function. The default is FALSE.
}
\examples{
x <- rnorm(10000)
plot(density(x),ylim=c(0,.5))

minor <- function() {
  plot(x,type="l",main="Trace",cex.main=.8,bty="n",cex.axis=.5) 
}

plotInPlot(minor,c(1,4,.4,.5))
plotInPlot(minor,"topright")
plotInPlot(minor,"bottomright")
plotInPlot(minor,"bottomleft")
plotInPlot(minor,"topleft")
}

