% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{plotPost}
\alias{plotPost}
\title{plot posterior}
\usage{
plotPost(x, ci = TRUE, stats = TRUE, trace = TRUE, dig = 3, cex.a = 1,
  col.area = "cornflowerblue", float = FALSE, ...)
}
\description{
plot posterior
}

