% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{plotPosts}
\alias{plotPosts}
\title{plot posteriors}
\usage{
plotPosts(M, cnames = colnames(M), stats = F, ...)
}
\description{
plot posteriors
}
\examples{
M <- matrix(rnorm(1000),ncol=4)
colnames(M) <- 1:4
plotPosts(M)
}

