\name{rcommon-package}
\alias{rcommon-package}
\alias{rcommon}
\docType{package}
\title{
\packageTitle{rcommon}
}
\description{
\packageDescription{rcommon}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rcommon}
\packageIndices{rcommon}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rcommon}

Maintainer: \packageMaintainer{rcommon}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
