% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgba.R
\name{rgba}
\alias{rgba}
\title{Add transparency to colors}
\usage{
rgba(col, alpha = 1)
}
\description{
Add transparency to colors
}
\examples{
plot(cbind(rnorm(300),rnorm(300)), col=rgba('steelblue', a=.5), pch=20, cex=3)
}
