% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorUnderCurve.R
\name{color.btwn.mat}
\alias{color.btwn.mat}
\title{Color area between ylo and yhi}
\usage{
color.btwn.mat(x, y, from = min(x), to = max(x), col.area = rgb(0, 0, 0,
  0.2))
}
\description{
Color area between ylo and yhi
}

